refactor: 온보딩 케이크 스타일 선택 그리드 별도의 컴포넌트로 분리
